Fix an abvious issue

+++ js/laz_rs_webassembly-worker.js
@@ -30,7 +30,7 @@ onmessage = async (event) => {
         min_z: result.header.min_z,
     };
 
-    const coordinates = new Float32Array(instance.memory.buffer, result.coordinates_pointer, result.coordinates_length);
+    const coordinates = new Float64Array(instance.memory.buffer, result.coordinates_pointer, result.coordinates_length);
     const intensities = new Uint16Array(instance.memory.buffer, result.intensity_pointer, result.intensity_length);
     const classifications = new Uint8Array(instance.memory.buffer, result.classification_pointer, result.classification_length);
